Migrate some unsafe helpers
